move functions from utils to encoding.rs
