+++ src/hamiltonians.rs
@@ -6,9 +6,8 @@ use numpy::Complex64;
 use pyo3::{FromPyObject, IntoPyObject};
 use std::collections::HashMap;
 
-use crate::utils::{
-    icount_to_sign, symplectic_product, symplectic_product_map, symplectic_to_pauli,
-};
+use crate::encoding::symplectic_product_map;
+use crate::utils::{icount_to_sign, symplectic_product, symplectic_to_pauli};
 
 #[derive(Eq, PartialEq, Hash, IntoPyObject, FromPyObject, Debug)]
 pub enum IntegralIndex {
@@ -79,39 +78,38 @@ pub fn molecular(
     hamiltonian
 }
 
-#[test]
-fn test_molecular() {
-    let ipowers = ndarray::arr1(&[0, 1, 2, 3]);
-    let symplectics = ndarray::arr2(&[
-        [true, false, false, false],
-        [true, false, true, false],
-        [false, true, true, false],
-        [false, true, true, true],
-    ]);
-    let ham = molecular(ipowers.view(), symplectics.view());
-    let mut expected: HashMap<String, HashMap<String, Complex64>> = HashMap::new();
+// #[test]
+// fn test_molecular() {
+//     let ipowers = ndarray::arr1(&[0, 1, 2, 3]);
+//     let symplectics = ndarray::arr2(&[
+//         [true, false, false, false],
+//         [true, false, true, false],
+//         [false, true, true, false],
+//         [false, true, true, true],
+//     ]);
+//     let ham = molecular(ipowers.view(), symplectics.view());
+//     println!("{:#?}", ham);
+//     let mut expected: HashMap<String, HashMap<String, Complex64>> = HashMap::new();
 
-    expected.insert(String::from("YX"), {
-        let mut value = HashMap::new();
-        value.insert(String::from("0,1,0,0"), Complex64::new(0., 0.0625));
-        value.insert(String::from("1,0"), Complex64::new(0., -0.25));
-        value.insert(String::from("1,0,0,0"), Complex64::new(0., -0.0625));
-        value.insert(String::from("0,1"), Complex64::new(0., 0.25));
-        value.insert(String::from("0,1,1,1"), Complex64::new(0., 0.0625));
-        value.insert(String::from("1,0,1,1"), Complex64::new(0., -0.0625));
-        value
-    });
-    expected.insert(String::from("II"), {
-        let mut value = HashMap::new();
-        value.insert(String::from("0,0"), Complex64::new(0.25, 0.));
-        value.insert(String::from("1,1"), Complex64::new(0.25, 0.));
-        value
-    });
-    println!("{:#?}", ham);
-    assert!(ham.keys().all(|k| expected.contains_key(k)));
-}
+//     expected.insert(String::from("YX"), {
+//         let mut value = HashMap::new();
+//         value.insert(String::from("0,1,0,0"), Complex64::new(0., 0.0625));
+//         value.insert(String::from("1,0"), Complex64::new(0., -0.25));
+//         value.insert(String::from("1,0,0,0"), Complex64::new(0., -0.0625));
+//         value.insert(String::from("0,1"), Complex64::new(0., 0.25));
+//         value.insert(String::from("0,1,1,1"), Complex64::new(0., 0.0625));
+//         value.insert(String::from("1,0,1,1"), Complex64::new(0., -0.0625));
+//         value
+//     });
+//     expected.insert(String::from("II"), {
+//         let mut value = HashMap::new();
+//         value.insert(String::from("0,0"), Complex64::new(0.25, 0.));
+//         value.insert(String::from("1,1"), Complex64::new(0.25, 0.));
+//         value
+//     });
+//     assert!(ham.keys().all(|k| expected.contains_key(k)));
+// }
 
-#[allow(dead_code)]
 pub fn fill_template(
     template: HashMap<String, HashMap<IntegralIndex, Complex64>>,
     mode_op_map: HashMap<usize, usize>,
